Moved file_ops out of the municipalities module
