+++ src/cobb_tracker/municipalities/cobb.py
@@ -1,4 +1,4 @@
-from cobb_tracker.municipalities import file_ops
+from cobb_tracker import file_ops
 from cobb_tracker.cobb_config import CobbConfig
 import requests
 import json
