+++ src/cobb_tracker/municipalities/marietta.py
@@ -3,7 +3,7 @@ import pathlib
 import re
 import os
 
-from cobb_tracker.municipalities import file_ops
+from cobb_tracker import file_ops
 from cobb_tracker.cobb_config import CobbConfig
 
 from threading import Thread
